test: Move registry methods to their own file.
